+++ src/IGarminTyp.h
@@ -195,7 +195,7 @@ class IGarminTyp : public QObject
         QTextCodec * getCodec(quint16 codepage);
         void decodeBitmap(QDataStream &in, QImage &img, int w, int h, int bpp);
         bool decodeBppAndBytes(int ncolors, int w, int flags, int& bpp, int& bytes);
-        bool decodeColorTable(QDataStream& in, QImage& img, int ncolors,  int maxcolor, bool hasAlpha);
+        bool decodeColorTable(QDataStream& in, QImage& img, quint8 ncolors,  int maxcolor, bool hasAlpha);
         format_e format;
 
         struct typ_section_t
